removed recoil as the movement system ahs goen in mroe of a fast paced system

+++ 2D top down/EventHandle.cpp
@@ -525,8 +525,8 @@ void EventHandle(SDL_Event& event)
 
 	SDL_Point facing = GetMapCoordFromCursor();
 
-	testPlayer->directionFacing.x = facing.x;
-	testPlayer->directionFacing.y = facing.y;
+	testPlayer->directionFacing.x = static_cast <float> (facing.x);
+	testPlayer->directionFacing.y = static_cast <float> (facing.y);
 
 	testPlayer->MovePlayerAccordingToInput();
